// frontend/src/app/models/registerUser.model.ts
export class registerUser {
  idUsersPk: number;
  username: string;
  email: string;
  name: string;
  lastName: string;
  typeUser: string;
  role: string;
  costPerHour: number;
  isActive: number;
  hoursPerWeek: number;
  idUnitFk: number;
  country: string;
  postalCode: string;
  bornDate: string;
  password: string;
  createdAt: Date;
}